refactor: File structure and test templates
